HW1/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
yfinance==0.2.28
beautifulsoup4==4.12.2